requirements.txt
altgraph==0.17.3
annotated-types==0.5.0
anyio==3.7.1
arabic-reshaper==3.0.0
click==8.1.4
colorama==0.4.6
EasyProcess==1.1
entrypoint2==1.1
fastapi==0.100.0
h11==0.14.0
idna==3.4
keyboard==0.13.5
mss==9.0.1
numpy==1.25.1
packaging==23.1
pefile==2023.2.7
Pillow==9.4.0
pydantic==2.0.2
pydantic_core==2.1.2
PyGetWindow==0.0.9
pyinstaller-hooks-contrib==2023.5
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
python-dotenv==1.0.0
pytweening==1.0.7
pywin32-ctypes==0.2.2
PyYAML==6.0
screeninfo==0.8.1
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
tksheet==5.4.1
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
websockets==11.0.3
win32-setctime==1.1.0
